// backend/src/modules/users/presentation/dto/admin-update-user.dto.ts
import { z } from 'zod';
import { UserRole, UserStatus, KycLevel } from '../../../../shared';

export const adminUpdateUserSchema = z.object({
  firstName: z.string().optional(),
  lastName: z.string().optional(),
  phone: z.string().optional().nullable(),
  role: z.nativeEnum(UserRole).optional(),
  status: z.nativeEnum(UserStatus).optional(),
  kycLevel: z.nativeEnum(KycLevel).optional(),
  emailVerified: z.boolean().optional(),
  password: z.string().min(6).optional(),
});

export type AdminUpdateUserDto = z.infer<typeof adminUpdateUserSchema>;